+++ src/misc.jl
@@ -1,8 +1,8 @@
-    language_codes = ("cn", "de", "el", "en", "es", "fr", "it", "pt", "ru")
+    const language_codes = ("cn", "de", "el", "en", "es", "fr", "it", "pt", "ru")
 
-    boolean( chance_of_getting_true=50)= rand(1:100) <= chance_of_getting_true
+    boolean(chance_of_getting_true=50)::Bool = rand(1:100) <= chance_of_getting_true
 
-    null_boolean()=Dict( 0=> "none", 1=> true, -1=> false)[rand(-1:1)]
+    null_boolean()::Union{Bool,String} = Dict{Int64,Any}( 0=> "none", 1=> true, -1=> false)[rand(-1:1)]
 
     #md5() = hexdigest("md5", "$(rand())")
 
@@ -10,28 +10,31 @@
 
     #sha256() = hexdigest("sha256", "$(rand())")
 
-    country_code()= executor(data["faker"]["address"]["country_code"])
-    language_code() = language_codes[rand(1:length(language_codes))]
-
-    locale()="$(language_code())_$(country_code())"
+    country_code()::String = rand(data["faker"]["address"]["country_code"])
+    language_code()::String = rand(language_codes)
+    locale(items::Tuple=(language_code(),"_",country_code()))::String = reduce(string, items )
 
     #=uuid4()==#
 
-    password( longitud=10, special_caracteres=true, digitos=true, upper_case=true, lower_case=true)=
-       ( caracteres = [];
-        for x=1:longitud
-             special_caracteres==true && (push!(caracteres, "$(rand('!':'+'))"))
-             digitos==true && (push!(caracteres, "$(rand(0:9))"))
-             upper_case==true && (push!(caracteres, "$(rand('A':'Z'))"))
-             lower_case==true && (push!(caracteres, "$(rand('a':'z'))"))
-        end;
+    function password( longitud=10, special_caracteres=true, digitos=true, upper_case=true, lower_case=true)::String
+        functions::Tuple = ();
 
-        for x=1:longitud
-            ini = rand(1:length(caracteres))
-            fin = rand(1:length(caracteres))
-            intermedio = caracteres[ini]
-            caracteres[ini] = caracteres[fin]
-            caracteres[fin] = intermedio
-        end;
+        if special_caracteres==true
+            functions = (functions..., random_special_caracteres )
+        end
 
-       join(caracteres,""); )
+        if digitos==true
+            functions = (functions..., random_digit )
+        end
+
+        if upper_case==true
+            functions = (functions..., random_uppercase_letter)
+        end
+
+        if lower_case==true
+            functions = (functions..., random_lowercase_letter)
+        end
+
+         return join(map( x -> rand(functions)(), 1:longitud),"")
+
+ end